Added customization and icon
Customization is now possible thanks to the Pebble Clay framework. This allows for the customization of the background colour, as well as the colour of the main circle. 

A little icon has also been added to make the app look less terrible on the Pebble menu. More professional is better!