Optimized Chalk, Emery display
The Pebble Round interface now actually fits on the round screen rather than having some text cut off. Emery also now has larger fonts to account for the increased pixel density.

There is also now an option to remember the duration of the last session and have that be the default breathing duration.