Close if no interaction when woken up
The app now closes if the user does not touch the watch within 30 seconds of it waking up and reminding them to breathe. This prevents the app from wasting battery from drawing and redrawing the PDC for hours on end if the user did not touch the watch (which was what it did before).