Fixed spacing
Tabs or spaces, that is the question. 

But which of them shall we say doth love the most? 

O, beware, my lord, of the third option;
it is the melding of both, which doth mock
The code it feeds on.

Out, damned spot! Out I say—One, two. Why, then, ’tis time to do ’t.

This commit fixes the inconsistent spacing that was in the code before. Now there are only tabs, and no spaces. (Liberty! Freedom! Tyranny is dead!)