Fixed vibration during HRV-set breathing session
The vibration used to be based on the chosen breathing duration, not the one that was dynamically calculated by using the heart rate. This has now been fixed.

Vibrations also previously did not play during quiet time. This was unintended behaviour and has now been fixed.

A heart icon has been added to the top text if heart rate variation is currently being used to calculate a breathing speed.