Removed unnecessary variable
To the tune of 🎶"The Itsy-Bitsy Spider"🎶: 

The itsy bitsy variable
Climbed into the codebase
Down came the keyboard clicks
And deleted the variable out

Up came the mouse cursor
And pushed this new commit
And the itsy bitsy variable
Just never came back again.

This commit (or the previous one, actually) also moves the position of the reminder window vibrations, because previously it would vibrate without having pushed the window. Awkward!