Fixed icon not working for App Store
Turns out for some weird reason I have to have the update icon marked as not targeted towards specific platforms in order for the app icon to be shown after an App Glance update. This is very strange and also not documented, so... 🤷 Not my fault, I say! Not my fault!